# Fix: override wildcard glob pattern (**) in resolveFilePathsFromPatterns