--- a/lib/plugins/package/lib/packageService.js
+++ b/lib/plugins/package/lib/packageService.js
@@ -207,7 +207,7 @@ module.exports = {
     params.include.forEach((pattern) => {
       patterns.push(pattern);
     });
-    return globby(params.include.concat(['**']), {
+    return globby(['**'].concat(params.include), {
       cwd: path.join(this.serverless.config.servicePath, prefix || ''),
       dot: true,
       silent: true,