--- a/lib/plugins/package/lib/packageService.js
+++ b/lib/plugins/package/lib/packageService.js
@@ -207,6 +207,9 @@ module.exports = {
     params.include.forEach((pattern) => {
       patterns.push(pattern);
     });
+    // NOTE: please keep this order of concatenating the include params
+    // rather than doing it the other way round!
+    // see https://github.com/serverless/serverless/pull/5825 for more information
     return globby(['**'].concat(params.include), {
       cwd: path.join(this.serverless.config.servicePath, prefix || ''),
       dot: true,