@svlapin 2019-02-13
Thanks for the review @pmuens :+1: 
Side note - actually I believe a better proper approach would be to take `options.exclude` into account while doing initial matching (like it was before https://github.com/serverless/serverless/pull/5574/files#diff-0333f308ebf44d187c45d0955bfe24c1R191) or rely on user-defined patterns only, but looks like it was intentionally avoided to speed up `excludeDevDependencies` somehow.